removed 'buttonhandler' requirement from sawmill.py.
This handler was an attempt to create a debounce in software for the button, however:
* The physical hardware button now has a capacitor to provide debounce
* The way the button is connected to other parts of the system (event) acts implicitly as a filter to prevent multiple, sequential starts.

Additionally:
button_example.py has been added for use with testing buttons.
