--- buttonhandler/__init__.py
@@ -1,1 +1,0 @@
-from .buttonhandler import ButtonHandler
